counter_V5.0/requirements.txt
gunicorn
gevent
absl-py==1.0.0
Flask==2.2.2
PyMySQL==1.0.2
webencodings==0.5.1
flask-mail
flask-login
redis
flask-login
scipy==1.8.1
numpy==1.21.5
matplotlib==3.2.2
opencv_python_headless
torch==1.7.1
torchvision==0.8.2
tqdm==4.63.0
Pillow==9.0.1
werkzeug==2.2.2
